사용자 정보 삭제 로직에 password 비교 추가

+++ src/user/application/user.service.ts
@@ -4,14 +4,14 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { User } from '../persistence/user.entity';
 import { UserRepository } from '../persistence/user.repository';
 import { UserDeleteDto } from '../web/dto/user-delete.dto';
+import * as bcrypt from 'bcryptjs';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(User)
-    private userRepository: UserRepository
-  ) {
-  }
+    private userRepository: UserRepository,
+  ) {}
 
   async updateUserName(
     id: number,
@@ -33,21 +33,22 @@ export class UserService {
     return user;
   }
 
-  async deleteUser(id: number, userDeleteDto: UserDeleteDto): Promise<void> {
-    const { username } = userDeleteDto;
+  async deleteUser(
+    id: number,
+    userDeleteDto: UserDeleteDto,
+    user: User,
+  ): Promise<void> {
+    const { username, password } = userDeleteDto;
 
-    const user = await this.userRepository.findOne({
-      where: { id: id, username: username },
-    });
-
-    let result;
-    if (user) {
-      result = await this.userRepository.delete(id);
+    if (
+      user &&
+      user.username === username &&
+      bcrypt.compare(password, user.password)
+    ) {
+      const result = await this.userRepository.delete(id);
+      console.log(`Result: ${result}`);
     } else {
-      //result.affected == 0
       throw new NotFoundException(`해당 Id를 가진 회원은 존재하지 않습니다.`);
     }
-
-    console.log('result', result);
   }
 }
